Added support for polygon groups in UmbralMesh.
git-svn-id: http://svn.purei.org/seventhumbral/trunk@288 b2dc461f-b7ed-4520-8422-8e9c2511a46d

+++ dataobjects/chunks/PgrpChunk.h
@@ -5,7 +5,7 @@
 class CPgrpChunk : public CBaseChunk
 {
 public:
-	typedef std::vector<unsigned int> TriangleArray;
+	typedef std::vector<uint16> TriangleArray;
 
 							CPgrpChunk();
 	virtual					~CPgrpChunk();
